Extract SQLite lock-contention helpers into sqlite_util.py (WS5 seam)
Move the SQLite busy/lock utilities out of engine.py into a dedicated
sqlite_util.py module: _is_sqlite_locked_error (exception-chain lock detection),
_sqlite_busy_timeout_ms (PRAGMA probe, internal), and the
_temporary_sqlite_busy_timeout context manager. engine.py imports the two it
calls (_is_sqlite_locked_error, _temporary_sqlite_busy_timeout); the probe
travels with them. The session-end timeout budget constant
(_SESSION_END_BUSY_TIMEOUT_MS) stays in engine.py as caller-side policy. Drops
the now-unused `contextmanager` and `typing.Iterator` imports from engine.py.

Behaviour-preserving refactor (WS5 engine decomposition). The moved code is
byte-identical to its pre-move form. Rollback: git revert this commit restores
the inline block; no data/state migration involved.
