+++ engine.py
@@ -17,9 +17,8 @@ import subprocess
 import threading
 import time
 import weakref
-from contextlib import contextmanager
 from pathlib import Path
-from typing import Any, Dict, Iterator, List, Optional
+from typing import Any, Dict, List, Optional
 
 from agent.context_engine import ContextEngine
 
@@ -97,6 +96,10 @@ from .message_content import (
     normalize_content_value,
     stored_text_content_for_pattern_matching,
     text_content_for_pattern_matching,
+)
+from .sqlite_util import (
+    _is_sqlite_locked_error,
+    _temporary_sqlite_busy_timeout,
 )
 from .store import MessageStore
 from .tokens import count_message_tokens, count_messages_tokens, count_tokens
@@ -326,45 +329,6 @@ def _git_runtime_identity(root: Path) -> dict[str, Any]:
         "plugin_git_dirty": None if dirty_output is None else bool(dirty_output),
         "plugin_git_remote": _git("config", "--get", "remote.origin.url") or "",
     }
-
-
-def _is_sqlite_locked_error(exc: BaseException) -> bool:
-    """Return True when an exception chain represents SQLite lock contention."""
-    seen: set[int] = set()
-    current: BaseException | None = exc
-    while current is not None and id(current) not in seen:
-        seen.add(id(current))
-        message = str(current).lower()
-        if isinstance(current, sqlite3.Error) and "locked" in message:
-            return True
-        current = current.__cause__ or current.__context__
-    return False
-
-
-def _sqlite_busy_timeout_ms(conn: sqlite3.Connection) -> int:
-    row = conn.execute("PRAGMA busy_timeout").fetchone()
-    return int(row[0]) if row and row[0] is not None else 0
-
-
-@contextmanager
-def _temporary_sqlite_busy_timeout(
-    connections: List[sqlite3.Connection | None],
-    timeout_ms: int,
-) -> Iterator[None]:
-    """Temporarily bound SQLite lock waits for gateway-critical paths."""
-    bounded_timeout = max(0, int(timeout_ms))
-    originals: list[tuple[sqlite3.Connection, int]] = []
-    for conn in connections:
-        if conn is None:
-            continue
-        original = _sqlite_busy_timeout_ms(conn)
-        conn.execute(f"PRAGMA busy_timeout={bounded_timeout}")
-        originals.append((conn, original))
-    try:
-        yield
-    finally:
-        for conn, original in reversed(originals):
-            conn.execute(f"PRAGMA busy_timeout={original}")
 
 
 _SYNTHETIC_ASSISTANT_NOISE = {
