+++ sqlite_util.py
@@ -1,0 +1,53 @@
+"""SQLite lock-contention helpers shared by the LCM engine.
+
+Isolated from ``engine.py`` (WS5 seam): detecting SQLite lock-contention error
+chains and temporarily bounding ``busy_timeout`` on gateway-critical paths are a
+cohesive, pure SQLite concern with no engine state. ``engine.py`` imports the
+two entry points it calls; the busy-timeout probe travels with them. Callers
+keep their own policy constants (for example the session-end timeout budget).
+"""
+
+from __future__ import annotations
+
+import sqlite3
+from contextlib import contextmanager
+from typing import Iterator, List
+
+
+def _is_sqlite_locked_error(exc: BaseException) -> bool:
+    """Return True when an exception chain represents SQLite lock contention."""
+    seen: set[int] = set()
+    current: BaseException | None = exc
+    while current is not None and id(current) not in seen:
+        seen.add(id(current))
+        message = str(current).lower()
+        if isinstance(current, sqlite3.Error) and "locked" in message:
+            return True
+        current = current.__cause__ or current.__context__
+    return False
+
+
+def _sqlite_busy_timeout_ms(conn: sqlite3.Connection) -> int:
+    row = conn.execute("PRAGMA busy_timeout").fetchone()
+    return int(row[0]) if row and row[0] is not None else 0
+
+
+@contextmanager
+def _temporary_sqlite_busy_timeout(
+    connections: List[sqlite3.Connection | None],
+    timeout_ms: int,
+) -> Iterator[None]:
+    """Temporarily bound SQLite lock waits for gateway-critical paths."""
+    bounded_timeout = max(0, int(timeout_ms))
+    originals: list[tuple[sqlite3.Connection, int]] = []
+    for conn in connections:
+        if conn is None:
+            continue
+        original = _sqlite_busy_timeout_ms(conn)
+        conn.execute(f"PRAGMA busy_timeout={bounded_timeout}")
+        originals.append((conn, original))
+    try:
+        yield
+    finally:
+        for conn, original in reversed(originals):
+            conn.execute(f"PRAGMA busy_timeout={original}")
